Changes are specially made to implement raycast
Raycast only works if  there is the fixture between the two given
points.
we want to raycast from the position of the cue ball to the start  of
the wall.
"self.pointobj" calculate the hit  point on the wall which is inline
with cueball position and user touch onscreen
so , bydefault projection of cueball will always hit the wall (if there
is no fixture between them).