Each body has two fixture.
what was need for a such change?
	Raycast is from the centre of cueball to the centre of the
projectball , so everytime during raycast half of the projectball image
will be on the other object. Because of this effect player cannot judge
where the cueball will hit.

How adding one more fixture helps ?
	The outer fixture size is larger than inner fixture, such that
centre of projectball touches the outer fixture and  the tip of project
ball touches the edge of inner fixture which is size of original circle.