Improvement in projection of cueball
Cueball should be able to hit other balls from the  its own centre ,
from its left side and from its right side.Raycast is from the centre of
cueball to the centre of the project ball. With only one fixture,
cueBall was never able to hit the balls from the sides. Becasue of the
outer fixutre the centre of project ball will always be on the outer
fixutre and image of project ball will just look like just touch the
ball and not getting inside the ball. This proper information about how
cueball is going to hit the colored balls in the game