intial Steps  before reloading the cueball.
billardBalls.lua:
To reload the cueball at the user selected position we need two bitmap
objects
1. cueballPositionPreview : draggable image on the table to preview the
position of cueball.
2.cueballNotPlaceableIndicator : to show the user cueball cannot be
placed here.

Deleted some print statements.
When the cueball is deleted , create and add the object of
ReloadTheCueball to billiardball class.

Once the cueball falls in to the pocket, we add and create the object of
reload class.
Reload class takes previewOfCueball obj and cannotplaceable object as an
input .