What happens inside the reload class .
Reload class takes the two bitmap objects from Billiard class.

Functions of reload class : check if the user selected place is inside
the table and not coincides with other balls on the table .

1.To select the place we have mouse Down  , mouse Up , mouse Move event
listener .
2.we create the physics body of the previewOfCueball bitmap object which
is filter sensor , which means it  can give the contact details , when
it touches other balls without using any physics property of colliding.
3. using mousejoint with previewOfCueball check for above valid
conditions and let the user to put the cueball.
3.Cueballnotplaceable is just use as an alert to show user has not
selected the proper place.