Role of World in the reloading of the Cueball and for the game .
Our world object is global throught the program. World Object updates
the postion of the body inside the onEnterFrame function using Word
Step() function .

When preview of cueball with physic body comes in contact with  bodies
of other balls , it is Begin contact and End contact of the world event
which lets the user know about.