Added  billiard stick to the game .
PositionOfStick.lua :
This class creates the stick object and its physics body .

Role of stick in the game :
You can move around the stick on mouse events .
Stick and projecting path will be in one one line and facing one
direction.
You can  pull the stick back to hit the cueball.
The distance by which it is pulled , is used to apply the force on the
cueball .
Update the position of stick everytime cueball position is changed .